Convert textures and remove duplicates
Please properly review.

There are several textures with same name, but different extension.
.mbm vs. .dds
.dds wins imho.

In other cases only a .mbm is present.
I made a .dds out of it.

Some .tga and .png I also converted to .dds

And then there are these weird cases with _psd.dds or _pds_NRM.dds, which are sometimes referenced and sometimes not ...
So I created copies off all these instances without _psd in the filename.