Increase IDCSP truss ejection force
Can @zorg2044 or someone fly another test flight to see how it goes? Also, remember to disable the RCS on the probes until they are separated. Was giving me a lot of issues. Once they are cleanly separated, switch back to them and give them a little jet burst to get them spinning.