Corona again #759
I'm pretty happy with this. It needs to be light to slow down and the overall mass now is closer to where it started, just less ablator and fuel. Tweaked the chutes.

Orient
Burn Retro
Orient prograde (optional, it self orients pretty well in the atmosphere)
Stage Chute - This jettisons the retro motor and arms the chute
Flaming reentry - it gets warm but hasn't blown up on me.
Chute auto deploys when safe
Jettison heat shield.

Letting it wobble freely in reentry helps a lot slowing down. It's most aerodynamic perfectly prograde.