Get GATV nose cone working. As far as I can tell, everything including shielding works.
Center (0,0,0 position) of the nose cone should be moved to where the CoM would be, OR closer to the attach node position for a more outward motion. Using MODEL position to adjust for now. ModuleDecouple applies force from the center of the decoupler part to the center of the part being decoupled. The way it's modeled now the center of the nose half is at the seam, which is directly inline with the docking port center, so the nose shoots straight ahead. Moving the center to the actual center of the nose cone half fixes that and the halves move away from each other.

Removed all MODULES from the Nose Cone and added ModuleBdbSymmetricalPart. It can manage the capped/uncapped just fine.

stackSymmetry set in Docking Port part. The docking port must be rotated 90 degrees for stackSymmetry to work. (KSP issue)

ModuleBdbSLAHelper, ModuleCargoBay, and ModuleDecouple's added to DockingPort.