Rotate Centaur / New boiloff beta A few random Centaur fixes.
The old boiloff code uses a halflife mechanism. At default 50% difficulty, when in sunlight, all tanks lost half their contents in 16 hours. That works out to about 4.5% per hour.

The new code makes better use of the information available in game to base boiloff rate on the amount of sunlight hitting the tank, and the amount of IR and reflected light coming from the body being orbited (I hope!!). In low orbit about a third of the radiation is coming from Kerbin (both bouncing sunlight and radiated heat). The angle of the tank to each body is taken into account, as well as if it is occluded by another part. For the same amount of incoming radiation, a larger amount of fuel will boil off slower than a lesser amount. Meaning the rate will increase as the tank is drained. This is the opposite of the half life method, where the rate continuously slowed.

Insulation and reflectivity (albedo) are properties of each part and can dramatically effect boiloff rate. Values of zero allow all radiation to pass, and a value of 1 allows none to pass. Both of these are simple multipliers to the radiation that hits the tank.

Q = radiation * Difficulty * (1 - insulation) * (1 - reflectivity)

The rates are currently higher than you're used to. This will be tuned in testing.