Scale Bluesmurff to support RSS (10.6x Kerbbin)
The current BLUESMURFF can balance well up to 6.4x, but fall fairly short for 10.6x that is the scale of the RealSolarSystem. This change should extend the balance toward 10.6x scale world, slightly less advantageous in smaller worlds like 6.4x and 2.5x, however, more consistent with the actual SMURFF mod that rebalance the other mods for 10.6x world.

There are going to have couple of errors for Vega_Upper_StageTank and Titan2_S1_UpperTank when SMURFF is installed, I think it is a bug on the SMURFF part since SMURFF isn't supposed to change the part anyway.