Scale BLUESMURFF to handle RSS (10.6X scale)
The original BLUESMURFF was designed to handle 6.4x world, but fell short on 10.6x world, e.g., this scale alone cannot lift Explorer-1 to the historical orbit (2550 x 358KM), so I further scaled this to the full SMURFF mod: dry mass reduction of 76% for liquid engine tank, 50% drymass reduction for SRB (40% in SMURFF, but we don't have that 40 ISP improvement).  I also need to boost some of the small kicker SRBs to reach reasonable thrust to payload ratio to so that Juno-I and Vanguard-I can reach orbit.